+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/entry/start.S7Atx6g3.js
@@ -1,0 +1,1 @@
+import{c as e,r as t}from"../chunks/C5TVoCB_.js";export{e as load_css,t as start};